fix: stabilize BLE sync protocol and implement synchronous auto-wipe
Firmware (oo-0.5.14):
- Standardized BLE protocol to 12-byte file list entries (Index + Time + Size) for stable targeting.
- Implemented 4-byte Little-Endian length prefixes for SD card audio frames to match app expectations.
- Added synchronous 'Auto-Wipe on Update' system: detects FW version mismatch in NVS and reformats SD card before boot pre-warming.
- Fixed 'Error ACK 8' (Index Out of Range) by marking files as deleted in cache instead of shifting indices mid-session.
- Fixed linker error for sys_rand32_get by adding random.h and enabling entropy generator.

App (0.3.4):
- Resolved sync deadlock by removing internal storage lock acquisition in OmiDeviceConnection.
- Implemented concurrency guard in DeviceProvider to prevent overlapping setup/sync loops.
- Updated SDCardWalSyncImpl to support 12-byte protocol and permissive file filtering (all durations visible).
- Disabled auto-connect and auto-resume features to provide manual control for debugging.
- Cleaned up architecture by removing unused Apple Watch discoverer and diagnostic logs.

+++ omi/firmware/omi/src/lib/core/transport.h
@@ -62,7 +62,7 @@ void put_current_connection(struct bt_conn *conn);
 extern uint32_t device_session_id;
 extern uint32_t segment_index;
 
-bool write_custom_packet_to_storage(uint8_t marker, uint8_t *data, uint8_t data_size);
+bool write_custom_packet_to_storage(uint32_t marker, uint8_t *data, uint32_t data_size);
 
 void transport_notify_button_state(uint8_t state);
